Missed one of the variables which should have been converted to off_t
git-svn-id: https://svn.code.sf.net/p/dansguardian/code/trunk@824 c6e6bffc-5bc6-4360-8735-ace3d95093af

+++ dansguardian/src/downloadmanagers/fancy.cpp
@@ -160,7 +160,7 @@ int fancydm::in(DataBuffer * d, Socket * sock, Socket * peersock, class HTTPHead
 	int rc;
 
 	off_t newsize;
-	int expectedsize = docheader->contentLength();
+	off_t expectedsize = docheader->contentLength();
 	off_t bytessec = 0;
 	off_t bytesgot = 0;
 	int percentcomplete = 0;
